services/scan-orchestrator/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
nats-py==2.7.2
redis[asyncio]==5.0.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
prometheus-fastapi-instrumentator==7.0.0
httpx==0.27.0
